Re-organize the editing files to prepare for adding future editors
