+++ apps/src/lib/levelbuilder/script-editor/LessonDescriptions.story.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import LessonDescriptions from './LessonDescriptions';
+import LessonDescriptions from '@cdo/apps/lib/levelbuilder/script-editor/LessonDescriptions';
 
 const currentDescriptions = [
   {
