+++ apps/src/lib/levelbuilder/script-editor/LessonGroups.jsx
@@ -1,10 +1,13 @@
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
-import color from '../../util/color';
-import {NEW_LEVEL_ID, addGroup} from './editorRedux';
-import NewLessonGroupInput from './NewLessonGroupInput';
-import LessonGroupCard from './LessonGroupCard';
+import color from '@cdo/apps/util/color';
+import {
+  NEW_LEVEL_ID,
+  addGroup
+} from '@cdo/apps/lib/levelbuilder/script-editor/editorRedux';
+import NewLessonGroupInput from '@cdo/apps/lib/levelbuilder/script-editor/NewLessonGroupInput';
+import LessonGroupCard from '@cdo/apps/lib/levelbuilder/script-editor/LessonGroupCard';
 
 const styles = {
   addGroupWithWarning: {
