+++ apps/src/lib/levelbuilder/script-editor/NewLessonGroupInput.jsx
