+++ apps/src/lib/levelbuilder/script-editor/NewScriptForm.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import RailsAuthenticityToken from '../util/RailsAuthenticityToken';
+import RailsAuthenticityToken from '@cdo/apps/lib/util/RailsAuthenticityToken';
 
 const buttonStyle = {
   marginLeft: 0,
