+++ apps/src/lib/levelbuilder/script-editor/ScriptEditor.jsx
@@ -1,17 +1,17 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import LessonGroups from './LessonGroups';
-import LessonDescriptions from './LessonDescriptions';
-import ScriptAnnouncementsEditor from './ScriptAnnouncementsEditor';
+import LessonGroups from '@cdo/apps/lib/levelbuilder/script-editor/LessonGroups';
+import LessonDescriptions from '@cdo/apps/lib/levelbuilder/script-editor/LessonDescriptions';
+import ScriptAnnouncementsEditor from '@cdo/apps/lib/levelbuilder/script-editor/ScriptAnnouncementsEditor';
 import $ from 'jquery';
 import ResourcesEditor from '@cdo/apps/templates/courseOverview/ResourcesEditor';
 import {resourceShape} from '@cdo/apps/templates/courseOverview/resourceType';
 import {announcementShape} from '@cdo/apps/code-studio/scriptAnnouncementsRedux';
-import VisibleAndPilotExperiment from './VisibleAndPilotExperiment';
+import VisibleAndPilotExperiment from '@cdo/apps/lib/levelbuilder/script-editor/VisibleAndPilotExperiment';
 import HelpTip from '@cdo/apps/lib/ui/HelpTip';
-import LessonExtrasEditor from './LessonExtrasEditor';
+import LessonExtrasEditor from '@cdo/apps/lib/levelbuilder/script-editor/LessonExtrasEditor';
 import color from '@cdo/apps/util/color';
-import MarkdownPreview from '@cdo/apps/lib/script-editor/MarkdownPreview';
+import MarkdownPreview from '@cdo/apps/lib/levelbuilder/script-editor/MarkdownPreview';
 
 const styles = {
   input: {
