+++ apps/src/sites/studio/pages/lessons/edit.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import getScriptData from '@cdo/apps/util/getScriptData';
-import LessonEditor from '@cdo/apps/lib/levelbuilder/LessonEditor';
+import LessonEditor from '@cdo/apps/lib/levelbuilder/lesson-editor/LessonEditor';
 
 $(document).ready(function() {
   const lessonData = getScriptData('lesson');
