+++ apps/src/sites/studio/pages/scripts/edit.js
@@ -6,8 +6,10 @@ import {Provider} from 'react-redux';
 import {getStore, registerReducers} from '@cdo/apps/redux';
 import isRtl from '@cdo/apps/code-studio/isRtlRedux';
 import getScriptData from '@cdo/apps/util/getScriptData';
-import reducers, {init} from '@cdo/apps/lib/script-editor/editorRedux';
-import ScriptEditor from '@cdo/apps/lib/script-editor/ScriptEditor';
+import reducers, {
+  init
+} from '@cdo/apps/lib/levelbuilder/script-editor/editorRedux';
+import ScriptEditor from '@cdo/apps/lib/levelbuilder/script-editor/ScriptEditor';
 import {valueOr} from '@cdo/apps/utils';
 
 export default function initPage(scriptEditorData) {
