+++ apps/src/sites/studio/pages/scripts/new.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import NewScriptForm from '@cdo/apps/lib/script-editor/NewScriptForm';
+import NewScriptForm from '@cdo/apps/lib/levelbuilder/script-editor/NewScriptForm';
 
 $(document).ready(() => {
   ReactDOM.render(<NewScriptForm />, document.getElementById('form'));
