+++ apps/src/templates/courseOverview/CourseEditor.js
@@ -6,7 +6,7 @@ import {resourceShape} from '@cdo/apps/templates/courseOverview/resourceType';
 import VisibleAndPilotExperiment from '../../lib/script-editor/VisibleAndPilotExperiment';
 import HelpTip from '@cdo/apps/lib/ui/HelpTip';
 import color from '@cdo/apps/util/color';
-import MarkdownPreview from '@cdo/apps/lib/script-editor/MarkdownPreview';
+import MarkdownPreview from '@cdo/apps/lib/levelbuilder/script-editor/MarkdownPreview';
 
 const styles = {
   input: {
