+++ apps/test/unit/lib/script-editor/LessonDescriptionsTest.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {shallow, mount} from 'enzyme';
 import {assert} from '../../../util/reconfiguredChai';
 import sinon from 'sinon';
-import LessonDescriptions from '@cdo/apps/lib/script-editor/LessonDescriptions';
+import LessonDescriptions from '@cdo/apps/lib/levelbuilder/script-editor/LessonDescriptions';
 
 const currentDescriptions = [
   {
