+++ apps/test/unit/lib/script-editor/LessonExtrasEditorTest.jsx
@@ -1,7 +1,7 @@
 import {expect} from 'chai';
 import React from 'react';
 import {mount} from 'enzyme';
-import LessonExtrasEditor from '@cdo/apps/lib/script-editor/LessonExtrasEditor';
+import LessonExtrasEditor from '@cdo/apps/lib/levelbuilder/script-editor/LessonExtrasEditor';
 
 const DEFAULT_PROPS = {
   projectWidgetVisible: false,
