+++ apps/test/unit/lib/script-editor/LessonGroupCardTest.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 import {expect} from '../../../util/reconfiguredChai';
-import {UnconnectedLessonGroupCard as LessonGroupCard} from '@cdo/apps/lib/script-editor/LessonGroupCard';
+import {UnconnectedLessonGroupCard as LessonGroupCard} from '@cdo/apps/lib/levelbuilder/script-editor/LessonGroupCard';
 
 const defaultProps = {
   lessonGroupsCount: 1,
