+++ apps/test/unit/lib/script-editor/RemoveLevelDialogTest.js
@@ -3,7 +3,7 @@ import {mount} from 'enzyme';
 import {expect} from '../../../util/reconfiguredChai';
 import sinon from 'sinon';
 
-import {UnconnectedRemoveLevelDialog as RemoveLevelDialog} from '@cdo/apps/lib/script-editor/RemoveLevelDialog';
+import {UnconnectedRemoveLevelDialog as RemoveLevelDialog} from '@cdo/apps/lib/levelbuilder/script-editor/RemoveLevelDialog';
 
 const levelKeyList = {
   22: 'Level Twenty Two'
