+++ apps/test/unit/lib/script-editor/ScriptAnnouncementsEditorTests.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 import {assert} from '../../../util/deprecatedChai';
-import ScriptAnnouncementsEditor from '@cdo/apps/lib/script-editor/ScriptAnnouncementsEditor';
+import ScriptAnnouncementsEditor from '@cdo/apps/lib/levelbuilder/script-editor/ScriptAnnouncementsEditor';
 
 const sampleAnnouncement = {
   notice: 'This course has recently been updated!',
